restructured data importer, separated parsers, BROKEN PARSERS

+++ mysql.py
@@ -13,7 +13,11 @@ ID(int), name(varchar), location(varchar)
 
 import MySQLdb
 
-def resetDB( db, defaultKeys = ["Name", "Email", "Address"] ):
+def resetDB( db, defaultKeys ):
+	if "location" not in [key.lower() for key in defaultKeys.keys()]
+		print "Can't create tables because there's no Location information..."
+		raise NameError("No Location key in defaultKeys")
+
 	c = db.cursor()
 	c.execute("DROP TABLE IF EXISTS tags")
 	c.execute("""CREATE TABLE tags (
@@ -22,12 +26,12 @@ def resetDB( db, defaultKeys = ["Name", "Email", "Address"] ):
 			PRIMARY KEY(id)
 			)""")
 	c.execute("DROP TABLE IF EXISTS items")
-	c.execute("""CREATE TABLE items (
-			id INT NOT NULL AUTO_INCREMENT,
-			name TINYTEXT NOT NULL,
-			location VARCHAR(20) NOT NULL,
-			PRIMARY KEY(id)
-			)""")
+	query = "".join([
+		"""CREATE TABLE items (
+			id INT NOT NULL AUTO_INCREMENT,""",
+			",\n".join([ " ".join([name, t]) for name, t in defaultKeys.values()]),
+			",\nPRIMARY KEY(id))"])
+	c.execute(query)
 	c.execute("DROP TABLE IF EXISTS relations")
 	c.execute("""CREATE TABLE relations (
 			id INT NOT NULL AUTO_INCREMENT,
@@ -36,14 +40,6 @@ def resetDB( db, defaultKeys = ["Name", "Email", "Address"] ):
 			PRIMARY KEY(id),
 			INDEX(tagID)
 			)""")
-	c.execute("DROP TABLE IF EXISTS info")
-	query = """CREATE TABLE info (
-			id INT NOT NULL AUTO_INCREMENT,
-			"""
-	query = "".join([query,
-			",\n".join([name + " " + t +" NOT NULL" for name, t in defaultKeys.values()]),
-			",\nPRIMARY KEY(id))"])
-	c.execute(query)
 	c.close()
 	db.commit()
 
@@ -51,11 +47,24 @@ def getConnection():
 	db = MySQLdb.connect(host='localhost', user='user', passwd='password', db='doed')
 	return db
 
-def insertItems( items, db ):
+def insertItems( items, defaultKeys, db ):
 	c = db.cursor()
+	# just getting keys in the correct order
+	# constructing a query in fairly complex manner
+	# This lad doesn't understand that either:
+	# (\_/)
+	# (O.o)
+	# (> <)
+
+	query = "".join(["INSERT INTO info(",
+			", ".join([key for _,(key,_) in sorted(defaultKeys.items())]),
+			") VALUES(",
+			", ".join( ["'%s'" for _ in range( len(defaultKeys.keys()) )] ),
+			")"])
+	vals = [tuple( [item[key] for key in sorted( defaultKeys.keys() )] ) \
+			for item in items ]
 	for item in items :
-		c.execute( "INSERT INTO items(name, location) VALUES(%s, %s)",
-			(item['Name'], item['Location']) )
+		c.executemany(query, vals ) )	
 	db.commit()
 	c.close()
 
@@ -80,14 +89,7 @@ def insertRel( rel, db_cursor):
 
 def insertInfo( items, defaultKeys, db ):
 	c = db.cursor()
-	query = "".join(["INSERT INTO info(",
-			", ".join([key for _,(key,_) in sorted(defaultKeys.items())]),
-			") VALUES(",
-			", ".join( ["%s" for _ in range( len(defaultKeys.keys()) )] ),
-			")"])
-
-	for item in items :
-		c.execute(query, tuple( [item[key] for key in sorted(defaultKeys.keys())] ) )
+	
 
 	c.close()
 def closeDB( db ):
